Add CodeMirrorContainer react component (WIP)
The idea is to create a component that will be the root of the React
tree for CodeMirror blocks.  If we didn't have to deal with changes,
the component could just consume the list of root nodes and the CM
markers and create react Portals to render each root node into the
correct place in the CM DOM.

Changes (and quarantine nodes) will complicate matters.

Right now no portals are created.  There's just an abundance of
console.log() calls.