Render root nodes through React Portals into the CodeMirror TextMarkers
When entering blocks mode, create TextMarkers in code mirror for each
root node.  Then pass the list of marks to CodeMirrorContainer as a
state update.  CMC will render each node into the TextMarker
replacement element using a portal.

The initial React tree will be
```
<CodeMirrorContainer>
  <React.Portal key="node-id of exprression1">
    <Expression>
      <!-- Nodes for expression1's children -->
    </Expression>
  </React.Portal>
  <React.Portal key="node-id of expression2">
    <Expression>
      <!-- Nodes for expression2's children -->
    </Expression>
  </React.Portal>
  <!-- etc -->
</CodeMirrorContainer>
```

TODO: successful changes should probably just recompute the marks for
the root nodes and push a state update with those and a cleared
quarantine flag.  I think that will drive react to reuse the existing
tree and just update any descendents that have a different node id.

TODO: quarantine should probably be a state change on the CMC too,
although I'm not sure if a portal will be the best way to hang the
change in the right place in the DOM